fixed agent

+++ src/components/adent_compnents/Agent_header.jsx
@@ -3,17 +3,41 @@ import { useNavigate } from "react-router-dom"
 import { useDispatch } from "react-redux"
 import { agent_logout } from "../../store/slices/agentAuthentication"
 import { toast } from "react-toastify"
+import axios from "axios"
+import { baseURL } from "../../baseUrls/Urls"
+import { useSelector } from "react-redux"
 
 function Agent_header() {
   const dispatch = useDispatch()
   const navigate = useNavigate()
+  const token = localStorage.getItem("access_token")
+  
 
-  const handleLogoutSubmit = () => {
-    localStorage.clear()
-    dispatch(agent_logout())
-    console.log("hhee")
-    navigate("/Agent_login_page")
-    toast.success("Logout successful. See you next time!")
+
+  const handleLogoutSubmit = async () => {
+    const agentId = localStorage.getItem("agent_uuid")
+    console.log(agentId)
+
+    if (!agentId) {
+      toast.error("No agent ID found. Please log in again.")
+      return
+    }
+
+    try {
+      const res = await axios.put(`${baseURL}/agent_auth/agent_logout/${agentId}`,{},{
+        headers: { Authorization: `Bearer ${token}`, },
+      })
+      
+      if (res.status === 200) {
+        localStorage.clear()
+        dispatch(agent_logout())
+        navigate("/Agent_login_page")
+        toast.success("Logout successful. See you next time!")
+      }
+    } catch (error) {
+      console.error("Logout failed:", error)
+      toast.error("Logout failed. Please try again.")
+    }
   }
 
   return (
